pyproject.toml
[project]
name = "fuckcodex"
version = "0.1.0"
description = "应用服务模块示例"
requires-python = ">=3.13"
dependencies = [
  "curl-cffi>=0.7.1",
  "google-api-python-client>=2.170.0",
  "google-auth-httplib2>=0.2.0",
  "google-auth-oauthlib>=1.2.2",
  "pydoll-python>=2.21.3",
  "rich>=14.3.3",
]
